+++ src/ProjectCostForecast.App/Presentation/ScheduleGridConverters.cs
@@ -2,7 +2,7 @@ using System.Globalization;
 using System.Windows;
 using System.Windows.Data;
 
-namespace ProjectCostForecast.App.Models;
+namespace ProjectCostForecast.App;
 
 public sealed class DateOnlyTextConverter : IValueConverter
 {